JSON output: cleaned up code per code-review.
