--- src/run_drakvuf.py
@@ -1,196 +1,0 @@
-#!/usr/bin/env python3
-
-##
-## Spawns DRAKVUF and parses its outbound messages via pipe.
-##
-
-## @TODO:
-## - add async support (poll on outpipe)
-## - improve ^C support
-## - remote breakpoints, verify correct handling of bad input
-
-
-import os
-import sys
-import signal
-
-import json
-import logging
-
-import subprocess # ???
-
-if not (sys.version_info.major >= 3 and sys.version_info.minor >= 6):
-    print("This script required Python 3.6+")
-    sys.exit(1)
-
-import curio
-import curio.subprocess # ???
-import asyncio
-import selectors
-
-
-proc = None # handle to DRAKVUF process
-
-def signal_handler(sig, frame):
-    global proc
-    logging.info("Handling signal %d: stopping DRAKVUF", sig)
-    if proc:
-        proc.terminate()
-    sys.exit(0)
-
-i = 0
-
-def get_event_nojson(outpipe):
-    global i
-    line = b''
-
-    while True:
-        line = outpipe.readline()
-        if not line:
-            break
-
-        i += 1
-        # try:
-        #     dline = line.decode("utf-8")
-        # except UnicodeDecodeError as e:
-        #     logging.warning("Failed to decode message %d [%s...]: %s", i, line[:40], e)
-        #     import pdb;pdb.set_trace()
-        #     break
-        yield line
-
-    # all done...
-    yield None
-
-def get_event(outpipe):
-    global i
-
-    line = b''
-    prevcol = -1 # column of previous error
-    while True:
-        #import pdb;pdb.set_trace()
-        line += outpipe.readline()
-        if not line:
-            break
-
-        i += 1
-        try:
-            dline = line.decode("utf-8")
-        except UnicodeDecodeError as e:
-            logging.warning("Failed to decode message %d [%s...]: %s", i, line[:40], e)
-            import pdb;pdb.set_trace()
-            break
-        try:
-            j = json.loads(dline)
-            line = b''
-            #print(j)
-        except json.decoder.JSONDecodeError as e:
-            logging.info("Failed to parse JSON from message %d [%s...]: %s", i, dline[:40], e)
-            if prevcol == e.colno:
-                logging.warning("Giving up an malformed JSON document:\n%s", dline)
-                line = b''
-                prevcol = -1
-            else:
-                prevcol = e.colno
-            import pdb;pdb.set_trace()
-            continue
-            # break
-        return j
-
-    # all done...
-    return None
-
-def parse_event(event):
-    global i
-
-    i += 1
-    if i % 1000 == 0:
-        print("Parsed {} events".format(i))
-
-def handle_output(outpipe):
-    json = get_event(outpipe)
-    print(json)
-    parse_event(json)
-
-def read_nop(outpipe):
-    line = outpipe.readline()
-    try:
-        dline = line.decode("utf-8")
-    except UnicodeDecodeError as e:
-        logging.warning("Failed to decode message %d [%s...]: %s", i, line[:40], e)
-        import pdb;pdb.set_trace()
-        return
-
-    print("Error: {}".format(dline))
-    return True
-
-def spawn_drakvuf(profile, domid):
-    global proc
-
-    fmt = 'json'
-    #fmt = 'kv'
-
-    cmd  = ['drakvuf', '-r', profile, '-d', domid, '-o', fmt]
-    #cmd.append('-v')
-
-    logging.info("Running %s", ' '.join(cmd))
-
-    # proc = curio.subprocess.Popen(cmd,
-    #                               stdout=curio.subprocess.PIPE,
-    #                               stderr=curio.subprocess.PIPE)
-
-    proc = subprocess.Popen(cmd,
-                            stdout=subprocess.PIPE,
-                            stderr=subprocess.PIPE)
-
-    stdout, stderr = proc.stdout, proc.stderr
-
-    #stdout.setblocking(False)
-    #stderr.setblocking(False)
-
-    sel = selectors.DefaultSelector()
-    sel.register(stdout, selectors.EVENT_READ, handle_output)
-    sel.register(stderr, selectors.EVENT_READ, read_nop)
-
-    process = True
-    while process:
-        import pdb;pdb.set_trace()
-        events = sel.select()
-        import pdb;pdb.set_trace()
-        for key, mask in events:
-            #import pdb;pdb.set_trace()
-            callback = key.data
-            val = callback(key.fileobj)
-            if val is None:
-                # fall out of outer loop
-                process = False
-                break
-
-    # #for e in get_event(proc.stdout):
-    # for e in get_event_nojson(proc.stdout):
-    #     if not e:
-    #         break
-
-    #     parse_event(e)
-
-    # No more input...
-    rc = proc.wait()
-    if rc:
-        logging.warn("Returned %d, stderr: %s", rc, proc.stderr.read())
-    proc.terminate()
-
-if __name__ == '__main__':
-    if len(sys.argv) != 3:
-        print("Usage: {} <rekall_profile> <domid>".format(sys.argv[0]))
-        print("Run as root (sudo)!")
-        sys.exit(1)
-
-    logging.basicConfig(level=logging.INFO)
-
-    (profile, domid) = sys.argv[1:]
-
-    signal.signal(signal.SIGHUP,  signal_handler)
-    signal.signal(signal.SIGINT,  signal_handler)
-    signal.signal(signal.SIGQUIT, signal_handler)
-    signal.signal(signal.SIGABRT, signal_handler)
-
-    spawn_drakvuf(profile, domid)
